requirements.txt
django==3.2.18
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow
django-storages==1.13.2
django-ckeditor==6.5.1
psycopg